+++ app/Config/Filters.php
@@ -19,6 +19,7 @@ class Filters extends BaseConfig
 		'csrf'     => CSRF::class,
 		'toolbar'  => DebugToolbar::class,
 		'honeypot' => Honeypot::class,
+		'AuthCheck' => \App\Filters\AuthCheckFilter::class,
 	];
 
 	/**