+++ app/Views/admin/template.php
@@ -242,7 +242,7 @@
 								</ul>
 							</li>
 							<li class="nav-item">
-								<a href="<?= base_url('admin/login/logout') ?>" class="nav-link">
+								<a href="<?= base_url('/login/logout') ?>" class="nav-link">
 									<i class="nav-icon fas fa-sign-out-alt"></i>
 									<p>
 										Sair